+++ utils/system-checks.sh
@@ -616,6 +616,11 @@ platform_sanity_checks() {
                # PrgEnv-pgi: currently neither GOOD nor BAD due to lack of testing
                # However, if logic above identified a bad version, we'll preserve that.
                unset COMPILER_GOOD
+            elif [[ $LMOD_FAMILY_PRGENV = PrgEnv-nvidia || $LMOD_FAMILY_PRGENV = PrgEnv-nvhpc ]]; then
+               # HPE Cray EX (Shasta) only validated for 21.9 and newer
+               if ! egrep ' +(21\.9|21\.1[0-9]|2[2-9]\.[0-9]+|[3-9][0-9]\.[0-9]+)-' <<<"$CXXVERS" 2>&1 >/dev/null ; then
+                  unset COMPILER_GOOD
+               fi
             fi
         elif echo "$CXXVERS" | egrep 'IBM XL'  > /dev/null ; then
             COMPILER_BAD=1
@@ -665,6 +670,19 @@ platform_sanity_checks() {
             COMPILER_BAD=1
         elif test x86_64 = "$ARCH" && echo "$CXXVERS" | egrep 'clang version ([4-9]\.|[1-9][0-9])' 2>&1 > /dev/null ; then
             COMPILER_GOOD=1
+            if [[ $LMOD_FAMILY_PRGENV = PrgEnv-aocc ]]; then
+               # HPE Cray EX (Shasta) only validated for aocc/3.1.0 and newer
+               # NOTE: 3.2 was end of 3.x series, but we'll accept up to 3.9 here
+               if ! egrep 'AOCC_(3\.[1-9]|[4-9]\.|[1-9][0-9]\.)' <<<"$CXXVERS" 2>&1 >/dev/null ; then
+                  unset COMPILER_GOOD
+               fi
+            elif [[ $LMOD_FAMILY_PRGENV = PrgEnv-amd ]]; then
+               # HPE Cray EX (Shasta) only validated for amd/4.2.0 and newer
+               # NOTE: 4.5 was end of 4.x series, but we'll accept up to 4.9 here
+               if ! egrep 'roc-(4\.[2-9]|[5-9]\.|[1-9][0-9]\.)' <<<"$CXXVERS" 2>&1 >/dev/null ; then
+                  unset COMPILER_GOOD
+               fi
+            fi
         elif test ppc64le = "$ARCH" && echo "$CXXVERS" | egrep 'clang version ([5-9]\.|[1-9][0-9])' 2>&1 > /dev/null ; then
 	    # Issue #236: ppc64le/clang support floor is 5.x. clang-4.x/ppc has correctness issues and is deliberately left "unvalidated"
             COMPILER_GOOD=1
@@ -736,6 +754,10 @@ We recommend one of the following C++ compilers (or any later versions where no 
                               ALCF's PrgEnv-llvm/4.0
            HPE Cray EX:       PrgEnv-gnu with gcc/10.3.0 environment module loaded
                               PrgEnv-cray with cce/12.0.0 environment module loaded
+                              PrgEnv-amd with amd/4.2.0 environment module loaded
+                              PrgEnv-aocc with aocc/3.1.0 environment module loaded
+                              PrgEnv-nvidia with nvidia/21.9 environment module loaded
+                              PrgEnv-nvhpc with nvhpc/21.9 environment module loaded
 EOF
         if test -n "$ARCH_BAD" ; then
             echo "ERROR: This version of UPC++ does not support the '$ARCH' architecture."