+++ ChangeLog.md
@@ -19,6 +19,10 @@ General features/enhancements: (see specification and programmer's guide for ful
 
 Infrastructure changes:
 
+* Support for additional compiler families on HPE Cray EX systems:
+    - AMD compilers via PrgEnv-amd and PrgEnv-aocc
+    - Nvidia compilers via PrgEnv-nvidia and PrgEnv-nvhpc
+    - See [INSTALL.md](INSTALL.md) for details such as minimum versions.
 * UPC++ library build now outputs a GASNet-EX configuration summary near the end
 * Integration with Berkeley UPC is now deprecated and may be removed in a future release.